[#3073] feat: Axon Framework 5 Test Fixture - different APIs based on when message
