+++ test/src/main/java/org/axonframework/test/fixture/AxonTestFixture.java
@@ -19,17 +19,12 @@ package org.axonframework.test.fixture;
 import org.axonframework.commandhandling.CommandBus;
 import org.axonframework.commandhandling.CommandMessage;
 import org.axonframework.commandhandling.CommandResultMessage;
-import org.axonframework.commandhandling.GenericCommandMessage;
 import org.axonframework.configuration.ApplicationConfigurer;
 import org.axonframework.configuration.NewConfiguration;
 import org.axonframework.eventhandling.EventMessage;
 import org.axonframework.eventhandling.EventSink;
-import org.axonframework.eventhandling.GenericEventMessage;
 import org.axonframework.messaging.Message;
 import org.axonframework.messaging.MessageTypeResolver;
-import org.axonframework.messaging.MetaData;
-import org.axonframework.messaging.unitofwork.AsyncUnitOfWork;
-import org.axonframework.messaging.unitofwork.ProcessingContext;
 import org.axonframework.test.aggregate.Reporter;
 import org.axonframework.test.matchers.FieldFilter;
 import org.axonframework.test.matchers.IgnoreField;
@@ -49,9 +44,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.CompletableFuture;
-import java.util.function.Consumer;
-import java.util.function.Function;
 import java.util.function.UnaryOperator;
 import java.util.stream.Stream;
 
@@ -117,12 +109,12 @@ public class AxonTestFixture implements AxonTestPhase.Setup {
 
     @Override
     public AxonTestPhase.Given given() {
-        return new Given(configuration, customization, commandBus, eventSink, messageTypeResolver);
+        return new AxonTestGiven(configuration, customization, commandBus, eventSink, messageTypeResolver);
     }
 
     @Override
     public AxonTestPhase.When when() {
-        return new When(configuration, customization, messageTypeResolver, commandBus, eventSink);
+        return new AxonTestWhen(configuration, customization, messageTypeResolver, commandBus, eventSink);
     }
 
     public record Customization(List<FieldFilter> fieldFilters) {
@@ -139,456 +131,10 @@ public class AxonTestFixture implements AxonTestPhase.Setup {
         public Customization registerIgnoredField(Class<?> declaringClass, String fieldName) {
             return registerFieldFilter(new IgnoreField(declaringClass, fieldName));
         }
-    }
-
-    /**
-     * Every given method invocation spawn new unit of work.
-     */
-    static class Given implements AxonTestPhase.Given {
-
-        private final NewConfiguration configuration;
-        private final Customization customization;
-        private final RecordingCommandBus commandBus;
-        private final RecordingEventSink eventSink;
-        private final MessageTypeResolver messageTypeResolver;
-        private final List<AsyncUnitOfWork> unitsOfWork = new ArrayList<>();
-
-        Given(
-                NewConfiguration configuration,
-                Customization customization,
-                RecordingCommandBus commandBus,
-                RecordingEventSink eventSink,
-                MessageTypeResolver messageTypeResolver
-        ) {
-            this.configuration = configuration;
-            this.customization = customization;
-            this.commandBus = commandBus;
-            this.eventSink = eventSink;
-            this.messageTypeResolver = messageTypeResolver;
-        }
 
         @Override
-        public AxonTestPhase.Given noPriorActivity() {
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Given event(Object payload, MetaData metaData) {
-            var eventMessage = toGenericEventMessage(payload, metaData);
-            return events(eventMessage);
-        }
-
-        private GenericEventMessage<Object> toGenericEventMessage(Object payload, MetaData metaData) {
-            var messageType = messageTypeResolver.resolve(payload);
-            return new GenericEventMessage<>(
-                    messageType,
-                    payload,
-                    metaData
-            );
-        }
-
-        @Override
-        public AxonTestPhase.Given events(List<?>... events) {
-            var messages = Arrays.stream(events)
-                                 .map(e -> e instanceof EventMessage<?> message
-                                         ? message
-                                         : toGenericEventMessage(e, MetaData.emptyInstance())
-                                 ).toArray(EventMessage<?>[]::new);
-            return events(messages);
-        }
-
-        @Override
-        public AxonTestPhase.Given events(EventMessage<?>... messages) {
-            inUnitOfWorkRunOnInvocation(processingContext -> eventSink.publish(processingContext,
-                                                                               messages));
-            return this;
-        }
-
-        private AsyncUnitOfWork inUnitOfWorkRunOnInvocation(Consumer<ProcessingContext> action) {
-            var unitOfWork = new AsyncUnitOfWork();
-            unitOfWork.runOnInvocation(action);
-            unitsOfWork.add(unitOfWork);
-            return unitOfWork;
-        }
-
-        private AsyncUnitOfWork inUnitOfWorkOnInvocation(Function<ProcessingContext, CompletableFuture<?>> action) {
-            var unitOfWork = new AsyncUnitOfWork();
-            unitOfWork.onInvocation(action);
-            unitsOfWork.add(unitOfWork);
-            return unitOfWork;
-        }
-
-        @Override
-        public AxonTestPhase.Given command(Object payload, MetaData metaData) {
-            var messageType = messageTypeResolver.resolve(payload);
-            var commandMessage = new GenericCommandMessage<>(
-                    messageType,
-                    payload,
-                    metaData
-            );
-            return commands(commandMessage);
-        }
-
-        @Override
-        public AxonTestPhase.Given commands(CommandMessage<?>... messages) {
-            inUnitOfWorkOnInvocation(processingContext -> {
-                CompletableFuture<? extends Message<?>> dispatchOneByOneFuture = CompletableFuture.completedFuture(null);
-                for (var message : messages) {
-                    var dispatchFuture = commandBus.dispatch(message, processingContext);
-                    dispatchOneByOneFuture = dispatchOneByOneFuture.thenCompose(m -> dispatchFuture);
-                }
-                return dispatchOneByOneFuture;
-            });
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.When when() {
-            for (var unitOfWork : unitsOfWork) {
-                awaitCompletion(unitOfWork.execute());
-            }
-            return new When(configuration, customization, messageTypeResolver, commandBus, eventSink);
-        }
-
-        private void awaitCompletion(CompletableFuture<?> completion) {
-            completion.join();
-        }
-    }
-
-
-    static class When implements AxonTestPhase.When {
-
-        private final NewConfiguration configuration;
-        private final Customization customization;
-        private final MessageTypeResolver messageTypeResolver;
-        private final RecordingCommandBus commandBus;
-        private final RecordingEventSink eventSink;
-        private final List<AsyncUnitOfWork> unitsOfWork = new ArrayList<>();
-
-        private Message<?> lastCommandResult;
-        private Throwable lastCommandException;
-
-        public When(
-                NewConfiguration configuration,
-                Customization customization,
-                MessageTypeResolver messageTypeResolver,
-                RecordingCommandBus commandBus,
-                RecordingEventSink eventSink
-        ) {
-            this.configuration = configuration;
-            this.customization = customization;
-            this.messageTypeResolver = messageTypeResolver;
-            this.commandBus = commandBus.reset();
-            this.eventSink = eventSink.reset();
-        }
-
-        @Override
-        public AxonTestPhase.When command(Object payload, MetaData metaData) {
-            var messageType = messageTypeResolver.resolve(payload);
-            var message = new GenericCommandMessage<>(messageType, payload, metaData);
-            inUnitOfWorkOnInvocation(processingContext ->
-                                             commandBus.dispatch(message, processingContext)
-                                                       .whenComplete((r, e) -> {
-                                                           if (e == null) {
-                                                               lastCommandResult = r;
-                                                               //lastCommandException = null;
-                                                           } else {
-                                                               //lastCommandResult = null;
-                                                               lastCommandException = e.getCause();
-                                                           }
-                                                       })
-            );
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.When event(Object payload) {
-            return AxonTestPhase.When.super.event(payload);
-        }
-
-        @Override
-        public AxonTestPhase.When event(Object payload, MetaData metaData) {
-            var eventMessage = toGenericEventMessage(payload, metaData);
-            return events(eventMessage);
-        }
-
-        private GenericEventMessage<Object> toGenericEventMessage(Object payload, MetaData metaData) {
-            var messageType = messageTypeResolver.resolve(payload);
-            return new GenericEventMessage<>(
-                    messageType,
-                    payload,
-                    metaData
-            );
-        }
-
-        @Override
-        public AxonTestPhase.When events(List<?>... events) {
-            var messages = Arrays.stream(events)
-                                 .map(e -> e instanceof EventMessage<?> message
-                                         ? message
-                                         : toGenericEventMessage(e, MetaData.emptyInstance())
-                                 ).toArray(EventMessage<?>[]::new);
-            return events(messages);
-        }
-
-        @Override
-        public AxonTestPhase.When events(EventMessage<?>... messages) {
-            inUnitOfWorkRunOnInvocation(processingContext -> eventSink.publish(processingContext,
-                                                                               messages));
-            return this;
-        }
-
-        private AsyncUnitOfWork inUnitOfWorkRunOnInvocation(Consumer<ProcessingContext> action) {
-            var unitOfWork = new AsyncUnitOfWork();
-            unitOfWork.runOnInvocation(action);
-            unitsOfWork.add(unitOfWork);
-            return unitOfWork;
-        }
-
-        private AsyncUnitOfWork inUnitOfWorkOnInvocation(Function<ProcessingContext, CompletableFuture<?>> action) {
-            var unitOfWork = new AsyncUnitOfWork();
-            unitOfWork.onInvocation(action);
-            unitsOfWork.add(unitOfWork);
-            return unitOfWork;
-        }
-
-
-        @Override
-        public AxonTestPhase.Then then() {
-            for (var unitOfWork : unitsOfWork) {
-                awaitCompletion(unitOfWork.execute());
-            }
-            return new Then(
-                    configuration,
-                    customization,
-                    messageTypeResolver,
-                    commandBus,
-                    eventSink,
-                    lastCommandResult,
-                    lastCommandException
-            );
-        }
-
-        private void awaitCompletion(CompletableFuture<?> completion) {
-            try {
-                completion.join();
-            } catch (Exception e) {
-                this.lastCommandResult = null;
-                this.lastCommandException = e;
-            }
-        }
-    }
-
-    static class Then implements AxonTestPhase.Then {
-
-        private final Reporter reporter = new Reporter();
-
-        private final NewConfiguration configuration;
-        private final Customization customization;
-        private final MessageTypeResolver messageTypeResolver;
-        private final RecordingCommandBus commandBus;
-        private final RecordingEventSink eventSink;
-
-        private final CommandValidator commandValidator;
-        private final Message<?> lastCommandResult;
-        private final Throwable lastCommandException;
-
-        public Then(
-                NewConfiguration configuration,
-                Customization customization,
-                MessageTypeResolver messageTypeResolver,
-                RecordingCommandBus commandBus,
-                RecordingEventSink eventSink,
-                Message<?> lastCommandResult,
-                Throwable lastCommandException
-        ) {
-            this.configuration = configuration;
-            this.customization = customization;
-            this.messageTypeResolver = messageTypeResolver;
-            this.commandBus = commandBus;
-            this.eventSink = eventSink;
-            this.lastCommandException = lastCommandException;
-            this.lastCommandResult = lastCommandResult;
-            this.commandValidator = new CommandValidator(commandBus::recordedCommands,
-                                                         commandBus::reset,
-                                                         new MatchAllFieldFilter(customization.fieldFilters));
-        }
-
-        @Override
-        public AxonTestPhase.Then events(Object... expectedEvents) {
-            var publishedEvents = eventSink.recorded();
-
-            if (expectedEvents.length != publishedEvents.size()) {
-                reporter.reportWrongEvent(publishedEvents, Arrays.asList(expectedEvents), lastCommandException);
-            }
-
-            Iterator<EventMessage<?>> iterator = publishedEvents.iterator();
-            for (Object expectedEvent : expectedEvents) {
-                EventMessage<?> actualEvent = iterator.next();
-                if (!verifyPayloadEquality(expectedEvent, actualEvent.getPayload())) {
-                    reporter.reportWrongEvent(publishedEvents, Arrays.asList(expectedEvents), lastCommandException);
-                }
-            }
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then events(EventMessage<?>... expectedEvents) {
-            this.events(Stream.of(expectedEvents).map(Message::getPayload).toArray());
-
-            var publishedEvents = eventSink.recorded();
-            Iterator<EventMessage<?>> iterator = publishedEvents.iterator();
-            for (EventMessage<?> expectedEvent : expectedEvents) {
-                EventMessage<?> actualEvent = iterator.next();
-                if (!verifyMetaDataEquality(expectedEvent.getPayloadType(),
-                                            expectedEvent.getMetaData(),
-                                            actualEvent.getMetaData())) {
-                    reporter.reportWrongEvent(publishedEvents, Arrays.asList(expectedEvents), lastCommandException);
-                }
-            }
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then events(Matcher<? extends List<? super EventMessage<?>>> matcher) {
-            var publishedEvents = eventSink.recorded();
-            if (!matcher.matches(publishedEvents)) {
-                final Description expectation = new StringDescription();
-                matcher.describeTo(expectation);
-
-                final Description mismatch = new StringDescription();
-                matcher.describeMismatch(publishedEvents, mismatch);
-
-                reporter.reportWrongEvent(publishedEvents, expectation, mismatch, lastCommandException);
-            }
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then commands(Object... expectedCommands) {
-            commandValidator.assertDispatchedEqualTo(expectedCommands);
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then commands(CommandMessage<?>... expectedCommands) {
-            commandValidator.assertDispatchedEqualTo(List.of(expectedCommands));
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then noCommands() {
-            commandValidator.assertDispatchedEqualTo(Matchers.noCommands());
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then success() {
-            return resultMessage(anything());
-        }
-
-        @Override
-        public AxonTestPhase.Then resultMessage(Matcher<? super CommandResultMessage<?>> matcher) {
-            if (matcher == null) {
-                return resultMessage(nullValue());
-            }
-            StringDescription expectedDescription = new StringDescription();
-            matcher.describeTo(expectedDescription);
-            if (lastCommandException != null) {
-                reporter.reportUnexpectedException(lastCommandException, expectedDescription);
-            } else if (!matcher.matches(lastCommandResult)) {
-                reporter.reportWrongResult(lastCommandResult, expectedDescription);
-            }
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then resultMessagePayload(Object expectedPayload) {
-            StringDescription expectedDescription = new StringDescription();
-            StringDescription actualDescription = new StringDescription();
-            PayloadMatcher<CommandResultMessage<?>> expectedMatcher =
-                    new PayloadMatcher<>(CoreMatchers.equalTo(expectedPayload));
-            expectedMatcher.describeTo(expectedDescription);
-            if (lastCommandException != null) {
-                reporter.reportUnexpectedException(lastCommandException, expectedDescription);
-            } else if (!verifyPayloadEquality(expectedPayload, lastCommandResult.getPayload())) {
-                PayloadMatcher<CommandResultMessage<?>> actualMatcher =
-                        new PayloadMatcher<>(CoreMatchers.equalTo(lastCommandResult.getPayload()));
-                actualMatcher.describeTo(actualDescription);
-                reporter.reportWrongResult(actualDescription, expectedDescription);
-            }
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then resultMessagePayloadMatching(Matcher<?> matcher) {
-            if (matcher == null) {
-                return resultMessagePayloadMatching(nullValue());
-            }
-            StringDescription expectedDescription = new StringDescription();
-            matcher.describeTo(expectedDescription);
-            if (lastCommandException != null) {
-                reporter.reportUnexpectedException(lastCommandException, expectedDescription);
-            } else if (!matcher.matches(lastCommandResult.getPayload())) {
-                reporter.reportWrongResult(lastCommandResult.getPayload(), expectedDescription);
-            }
-            return this;
-        }
-
-        @Override
-        public AxonTestPhase.Then exception(Class<? extends Throwable> expectedException) {
-            return exception(instanceOf(expectedException));
-        }
-
-        @Override
-        public AxonTestPhase.Then exception(Matcher<?> matcher) {
-            StringDescription description = new StringDescription();
-            matcher.describeTo(description);
-            if (lastCommandException == null) {
-                reporter.reportUnexpectedReturnValue(lastCommandResult.getPayload(), description);
-            }
-            if (!matcher.matches(lastCommandException)) {
-                reporter.reportWrongException(lastCommandException, description);
-            }
-            return this;
-        }
-
-        @SuppressWarnings("BooleanMethodIsAlwaysInverted")
-        private boolean verifyPayloadEquality(Object expectedPayload, Object actualPayload) {
-            if (Objects.equals(expectedPayload, actualPayload)) {
-                return true;
-            }
-            if (expectedPayload != null && actualPayload == null) {
-                return false;
-            }
-            if (expectedPayload == null) {
-                return false;
-            }
-            if (!expectedPayload.getClass().equals(actualPayload.getClass())) {
-                return false;
-            }
-            Matcher<Object> matcher = deepEquals(expectedPayload, new MatchAllFieldFilter(customization.fieldFilters));
-            if (!matcher.matches(actualPayload)) {
-                reporter.reportDifferentPayloads(expectedPayload.getClass(), actualPayload, expectedPayload);
-            }
-            return true;
-        }
-
-        @SuppressWarnings("BooleanMethodIsAlwaysInverted")
-        private boolean verifyMetaDataEquality(Class<?> eventType, Map<String, Object> expectedMetaData,
-                                               Map<String, Object> actualMetaData) {
-            MapEntryMatcher matcher = new MapEntryMatcher(expectedMetaData);
-            if (!matcher.matches(actualMetaData)) {
-                reporter.reportDifferentMetaData(eventType,
-                                                 matcher.getMissingEntries(),
-                                                 matcher.getAdditionalEntries());
-            }
-            return true;
-        }
-
-        @Override
-        public AxonTestPhase.Setup and() {
-            return AxonTestFixture.with(configuration, c -> customization);
+        public List<FieldFilter> fieldFilters() {
+            return fieldFilters;
         }
     }
 }
